requirements.txt
Flask==3.0.3
playwright==1.55.0
python-dotenv==1.0.1
supabase==2.4.2
